it jankly works tested and confirmed
there are a couple of issues that need to be addressed (that I didn't catch during earlier PRs that I might have to fix now): The dashboard widget is messing with wheelEvents and when you hover over a opened dropdown, it tries to scroll the dashboard instead of scroll through the dropdown. This is the biggest pain but will need to be addressed. Then there are some minor things such as when the layout auto expands, I don't think the boudning box is updated so the border is a little whack. Also, I havne't actually sent a CAN message yet :skull: also it would be nice to tab and shift-tab through the fields (OH RIGHT THATS WHY I NEEDED OT ABSTRACT IT AWAYT CAUSE QCOMBOBOXES ALSO NEED THE SIGNAL). Anyways, pain.