LFG EVENTS NOW PROPOGATE TO DASHBOARD ITEMS
thanks jack what I needed was just a break. proxy.setFocusPolicy(Qt.NoFocus) was the issue because by default, QGraphicsProxyWidget (the thing that embeds QWidgets onto the QGraphicsScene) should pass events, but if you set the focus policy to cant focus, the events dont get sent to it, which means it cant forward events to the children because there are no events to forward. HOLYYY OKAY I CAN SEE THE END OF THE TUNNEL WE MIGHT ACTUALLY BE ABLE TO FINISH THIS LFGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGG